Release v0.8: sky bilinear fix, shader packs, shadow mapping, clang-tidy pass
## Sky rendering
- Fix pixelated sky: d3dStartBufferG() forced uAlphaTest=true for all geometry
  buffer draws, engaging the foliage texel-snap UV path in basic.frag even for
  the sky texture. Added noSnapUV parameter to d3dEndBufferG(); sky passes TRUE
  to draw with uAlphaTest=false and genuine bilinear filtering. Matches original
  D3D smooth-sky mode.

## Shadow mapping (Phase 2.1 complete)
- Hardware PCF (sampler2DShadow + GL_LINEAR + GL_COMPARE_REF_TO_TEXTURE):
  3x3 kernel tap; 60 GU NDC bias; no self-shadowing acne on flat/mild slopes.
- Transparent geometry suppression: SetRenderStates(zWrite=false) during shadow
  depth pass sets m_shadowGeomSuppressed; draw calls skip while suppressed,
  eliminating spurious large underwater shadows that moved with the player.
- Sky plane excluded from shadow pass (sz=0.0001 reconstructs world pos ~160k
  GU ahead; GL_DEPTH_CLAMP would blacken the shadow map near sun elevation).
- World-space object/character shadows via SubmitWorldSpaceShadowTriangle.
- Dynamic frustum range tied to ctViewR.

## Shader pack system
- Modular post-process shader system: shaderpacks/ directory, pack.json format,
  ShaderPack/ShaderPackManager infrastructure.
- Default pack: ACES filmic tone mapping.
- Example pack: Reinhard tone mapping, color grade, bloom, vignette.
- Documentation: SHADER_PACKS.md (modder guide), SHADER_DEVELOPMENT_NOTES.md.

## clang-tidy pass (non-stylistic issues)
- Dead stores removed: renderd3d.cpp (C, zs, d, BL x2, sh initial), XR.cpp (rx x2).
- Integer division in float context: Characters.cpp x5, Hunt2.cpp x5,
  renderd3d.cpp x3 — added .0f suffixes.
- Duplicate branch bodies collapsed: Characters.cpp x5 AI state sites,
  Characters.cpp case AI_PTERA/DIMOR, XR.cpp switch, renderd3d.cpp,
  Materials.cpp (mr ? 1.0f : 1.0f).
- Insecure strcpy replaced with strncpy+NUL: Resources.cpp x6, Game.cpp,
  Interface.cpp.
- Stream state guards added in Resources.cpp fgets loops.
- OpenXR enum zero-init false positives annotated NOLINT (standard XR idiom).

+++ TextureOverrides.cpp
@@ -147,7 +147,7 @@ bool RegisterFromFile(void* key, const char* imagePath)
     stbi_uc* data = stbi_load(resolved.c_str(), &w, &h, &comp, 4);
     if (!data) return false;
 
-    uint32_t* buf = new uint32_t[w * h];
+    uint32_t* buf = new uint32_t[(size_t)w * h]; // SOURCEPORT: widen before multiply to avoid int overflow for large textures
     std::memcpy(buf, data, (size_t)w * h * 4);
     stbi_image_free(data);
